moved parser into separate module
